+++ packages/react/src/types/instance.ts
@@ -228,16 +228,16 @@ export type GeneralHelpers<NodeType extends Node = Node, EdgeType extends Edge =
     nodeId: string;
     handleId?: string | null;
   }) => NodeConnection[];
-  // /**
-  //  * Fits the view.
-  //  *
-  //  * @param options.padding - optional padding
-  //  * @param options.includeHiddenNodes - optional includeHiddenNodes
-  //  * @param options.minZoom - optional minZoom
-  //  * @param options.maxZoom - optional maxZoom
-  //  * @param options.duration - optional duration. If set, a transition will be applied
-  //  * @param options.nodes - optional nodes to fit the view to
-  //  */
+  /**
+   * Fits the view based on the passed params. By default it fits the view to all nodes.
+   *
+   * @param options.padding - optional padding
+   * @param options.nodes - optional nodes to fit the view to
+   * @param options.minZoom - optional minZoom
+   * @param options.maxZoom - optional maxZoom
+   * @param options.duration - optional duration. If set, a transition will be applied
+   * @param options.includeHiddenNodes - optional includeHiddenNodes
+   */
   fitView: FitView<NodeType>;
 };
 /**