+++ packages/svelte/src/lib/actions/zoom/index.ts
@@ -35,6 +35,7 @@ type ZoomParams = {
   userSelectionActive: boolean;
   lib: string;
   paneClickDistance: number;
+  selectionOnDrag?: boolean;
   onTransformChange: (transform: Transform) => void;
   onDraggingChange: (dragging: boolean) => void;
   connectionInProgress: boolean;
@@ -49,7 +50,6 @@ export default function zoom(domNode: Element, params: ZoomParams) {
     onPanZoom,
     onPanZoomEnd,
     translateExtent,
-    paneClickDistance,
     setPanZoomInstance,
     onDraggingChange,
     onTransformChange
@@ -61,7 +61,6 @@ export default function zoom(domNode: Element, params: ZoomParams) {
     maxZoom,
     translateExtent,
     viewport: initialViewport,
-    paneClickDistance,
     onPanZoom,
     onPanZoomStart,
     onPanZoomEnd,