+++ packages/svelte/src/lib/plugins/EdgeToolbar/types.ts
@@ -1,0 +1,8 @@
+import type { EdgeToolbarBaseProps } from '@xyflow/system';
+import type { Snippet } from 'svelte';
+import type { HTMLAttributes } from 'svelte/elements';
+
+export type EdgeToolbarProps = Omit<EdgeToolbarBaseProps, 'edgeId'> & {
+  selectEdgeOnClick?: boolean;
+  children?: Snippet;
+} & HTMLAttributes<HTMLDivElement>;