+++ packages/svelte/src/lib/store/index.ts
@@ -191,10 +191,6 @@ export function createStore<NodeType extends Node = Node, EdgeType extends Edge 
     }
   }
 
-  function setPaneClickDistance(distance: number) {
-    store.panZoom?.setClickDistance(distance);
-  }
-
   function deselect<T extends Node | Edge>(
     elements: T[],
     elementsToDeselect: Set<string> | null = null
@@ -235,12 +231,8 @@ export function createStore<NodeType extends Node = Node, EdgeType extends Edge 
       const nodeWillBeSelected = ids.includes(node.id);
       const selected = isMultiSelection ? node.selected || nodeWillBeSelected : nodeWillBeSelected;
 
-      if (node.selected !== selected) {
-        // we need to mutate the node here in order to have the correct selected state in the drag handler
-        const internalNode = store.nodeLookup.get(node.id);
-        if (internalNode) internalNode.selected = selected;
-        node.selected = selected;
-        return { ...node };
+      if (!!node.selected !== selected) {
+        return { ...node, selected };
       }
       return node;
     });
@@ -257,7 +249,7 @@ export function createStore<NodeType extends Node = Node, EdgeType extends Edge 
       const edgeWillBeSelected = ids.includes(edge.id);
       const selected = isMultiSelection ? edge.selected || edgeWillBeSelected : edgeWillBeSelected;
 
-      if (edge.selected !== selected) {
+      if (!!edge.selected !== selected) {
         return { ...edge, selected };
       }
       return edge;
@@ -397,7 +389,6 @@ export function createStore<NodeType extends Node = Node, EdgeType extends Edge 
     setMinZoom,
     setMaxZoom,
     setTranslateExtent,
-    setPaneClickDistance,
     unselectNodesAndEdges,
     addSelectedNodes,
     addSelectedEdges,